+++ src/core/settings.js
@@ -31,7 +31,8 @@ export class Settings {
     this._locations = {
       'api': 'https://apps.ionic.io',
       'push': 'https://push.ionic.io',
-      'analytics': 'https://analytics.ionic.io'
+      'analytics': 'https://analytics.ionic.io',
+      'deploy': 'https://apps.ionic.io'
     };
     this._devLocations = this.get('dev_locations');
     if (!this._devLocations) { this._devLocations = {}; }