+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.0
+
+* (core) removing need to call `Ionic.io();` manually
+* (deploy) update() can now defer reloading the app
+* 
+
 ## 0.1.0
 
 * web client introduction